ScreenCapture 関連を Win32Wrap に移動
