+++ WinRealCaptureConsole/WinRealCaptureConsole.csproj
@@ -47,7 +47,6 @@
   <ItemGroup>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ScreenCapture.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
